+++ transcendence/base/static/js/api.js
@@ -17,21 +17,15 @@ const apiHandler = {
     try {
       let response = await fetch(fullUrl, { ...options, headers });
 
-      if (response.status === 401 && refreshToken) {
-        const newTokens = await refreshAccessToken(refreshToken);
-        storeTokens(newTokens);
-
-        headers.Authorization = `Bearer ${newTokens.access}`;
-        response = await fetch(fullUrl, { ...options, headers });
-      }
-
-      const data = await response.json();
-      if (!response.ok) {
+      if (!response.ok || response.status === 401) {
+        console.log('crisse');
         localStorage.removeItem("accessToken");
         localStorage.removeItem("refreshToken");
         redirectToLogin();
         throw new Error(data.detail || "Something went wrong");
       }
+
+      const data = await response.json();
 
       return data;
     } catch (error) {
